build: Setup project and add typescript support
- Add a babelrc config file for typescript
- Installed missing packages for typescript support
- Add tsconfig file
- Change the js files to ts files

feature: Add a headline and two textboxes

styles: Adjust the styles for the headlines and the textboxes

test: Fixed some test and add some todo tests
