+++ src/app/App.css
@@ -1,0 +1,54 @@
+.App {
+  text-align: center;
+  font-family: Arial, Helvetica, sans-serif;
+}
+
+.App-header {
+  background-color: #282c34;
+  height: 25vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  font-size: calc(10px + 2vmin);
+  color: white;
+}
+
+.App-main {
+  background-color: #404040;
+  height: 75vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: flex-start;
+}
+
+.App-textfield {
+  width: 80vw;
+  font-family: Arial, Helvetica, sans-serif;
+  font-size: medium;
+  font-weight: 200;
+  padding: 1vw;
+}
+
+.App-textarea {
+  width: 80vw;
+  font-family: Arial, Helvetica, sans-serif;
+  font-size: medium;
+  height: 30vh;
+  font-weight: 200;
+  resize: none;
+  margin-top: 2vh;
+  padding: 1vw;
+}
+
+
+
+@keyframes App-logo-spin {
+  from {
+    transform: rotate(0deg);
+  }
+  to {
+    transform: rotate(360deg);
+  }
+}
